fix(atelier): complete the render_ir relocation imports (#1760)
PR #1817 committed the `codegen.rs` -> `render_ir.rs` move but, due to a
`git add` pathspec error, dropped the accompanying edits — leaving canary unable
to compile: `vize_relief` still declared `pub mod codegen` (the file was gone),
and `render_ir.rs` kept its old `use super::` imports (which resolve to
vize_relief modules that don't exist in vize_atelier_core).

This applies the missing edits:
- `render_ir.rs` imports the borrowed Relief types from `vize_relief` instead of
  the now-nonexistent `super::` modules, and references `TextNode` /
  `InterpolationNode` directly.
- `vize_relief/relief.rs` drops `pub mod codegen` / `pub use codegen::*`.
- `vize_atelier_core/lib.rs` re-exports the moved IR types from `render_ir` and
  keeps the genuine Relief types re-exported from `vize_relief`.
- `relief/tests.rs` drops the codegen-type constructor tests (they moved to a
  `vize_atelier_core` integration test, `tests/render_ir.rs`).

Restores a compiling, byte-equivalent canary.

BREAKING CHANGE: vize_relief no longer exposes the render/codegen IR types
(JsChildNode, VNodeCall, etc.); they live in vize_atelier_core (re-exported).

Verified: cargo build --workspace; vize_atelier_core (261 + render_ir 4),
vize_atelier_dom, vize_atelier_sfc (528), vize_atelier_ssr (76),
vize_relief (49); clippy clean.
