+++ crates/vize_relief/src/relief.rs
@@ -4,7 +4,6 @@
 //! target. All nodes are allocated in a bumpalo arena for efficient memory
 //! management and zero-copy transfer to JavaScript.
 
-pub mod codegen;
 pub mod control_flow;
 pub mod core;
 pub mod elements;
@@ -14,7 +13,6 @@ pub mod nodes;
 #[cfg(test)]
 mod tests;
 
-pub use codegen::*;
 pub use control_flow::*;
 pub use core::*;
 pub use elements::*;
